feature(backend) move enums and types to dedicated files and clean up code & tests
