+++ backend/src/main/scala/com/just/donate/Server.scala
@@ -8,11 +8,9 @@ import com.just.donate.api.OrganisationRoute.organisationApi
 import com.just.donate.api.PaypalRoute.paypalRoute
 import com.just.donate.api.TransferRoute.transferRoute
 import com.just.donate.api.WithdrawalRoute.withdrawalRoute
-import com.just.donate.config.AppEnvironment
-import com.just.donate.config.{ AppConfig, Config }
+import com.just.donate.config.{AppConfig, AppEnvironment, Config}
 import com.just.donate.db.PaypalRepository
-import com.just.donate.notify.DevEmailService
-import com.just.donate.notify.{ EmailService, IEmailService }
+import com.just.donate.notify.{DevEmailService, EmailService, IEmailService}
 import com.just.donate.store.FileStore
 import org.http4s.*
 import org.http4s.ember.server.*
