+++ backend/src/main/scala/com/just/donate/api/DonationRoute.scala
@@ -3,7 +3,8 @@ package com.just.donate.api
 import cats.effect.*
 import cats.implicits.*
 import com.just.donate.config.Config
-import com.just.donate.models.{ Donation, DonationError, Donor, Organisation }
+import com.just.donate.models.errors.{DonationError, TransferError, WithdrawError}
+import com.just.donate.models.{Donation, Donor, Organisation}
 import com.just.donate.notify.IEmailService
 import com.just.donate.store.Store
 import com.just.donate.utils.RouteUtils.loadAndSaveOrganisationOps
@@ -15,14 +16,6 @@ import org.http4s.circe.CirceSensitiveDataEntityDecoder.circeEntityDecoder
 import org.http4s.dsl.io.*
 
 object DonationRoute:
-  private val emailTemplate: (String, String, String) => String = (linkWithId, id, link) =>
-    f"""Thank you for your donation, to track your progress visit
-       |${linkWithId}
-       |or enter your tracking id
-       |${id}
-       |on our tracking page
-       |${link}""".stripMargin
-
   val donationRoute: (Store, Config, IEmailService) => HttpRoutes[IO] = (store, config, emailService) =>
     HttpRoutes.of[IO]:
 
@@ -44,6 +37,13 @@ object DonationRoute:
         yield response).handleErrorWith {
           case e: InvalidMessageBodyFailure => BadRequest(e.getMessage)
         }
+  private val emailTemplate: (String, String, String) => String = (linkWithId, id, link) =>
+    f"""Thank you for your donation, to track your progress visit
+       |${linkWithId}
+       |or enter your tracking id
+       |${id}
+       |on our tracking page
+       |${link}""".stripMargin
 
   private def organisationMapper(requestDonation: RequestDonation, accountName: String)(
     org: Organisation
