+++ backend/src/main/scala/com/just/donate/api/NotificationRoute.scala
@@ -12,15 +12,16 @@ import org.http4s.dsl.io.*
 
 object NotificationRoute:
 
-  val notificationRoute: Config => HttpRoutes[IO] = (config: Config) => HttpRoutes.of[IO]:
+  val notificationRoute: Config => HttpRoutes[IO] = (config: Config) =>
+    HttpRoutes.of[IO]:
 
-    case req @ POST -> Root / donor =>
-      for
-        notification <- req.attemptAs[NotificationRequest].value
-        _ <- notification match
-          case Right(NotificationRequest(message)) => new EmailService(config).sendEmail(donor, message)
-          case Left(_)                             => BadRequest()
-        response <- Ok("Notification sent")
-      yield response
+      case req @ POST -> Root / donor =>
+        for
+          notification <- req.attemptAs[NotificationRequest].value
+          _ <- notification match
+            case Right(NotificationRequest(message)) => new EmailService(config).sendEmail(donor, message)
+            case Left(_)                             => BadRequest()
+          response <- Ok("Notification sent")
+        yield response
 
   private case class NotificationRequest(message: String)
