+++ backend/src/main/scala/com/just/donate/api/TransferRoute.scala
@@ -3,6 +3,8 @@ package com.just.donate.api
 import cats.effect.*
 import cats.syntax.all.*
 import com.just.donate.config.Config
+import com.just.donate.models.Organisation
+import com.just.donate.models.errors.WithdrawError
 import com.just.donate.notify.IEmailService
 import com.just.donate.store.Store
 import com.just.donate.utils.RouteUtils.loadAndSaveOrganisationOps
@@ -12,7 +14,6 @@ import org.http4s.*
 import org.http4s.circe.*
 import org.http4s.circe.CirceSensitiveDataEntityDecoder.circeEntityDecoder
 import org.http4s.dsl.io.*
-import com.just.donate.models.{ Organisation, TransferError }
 
 object TransferRoute:
