+++ backend/src/main/scala/com/just/donate/config/Config.scala
@@ -8,4 +8,3 @@ trait Config:
   val mailSmtpPort: Int
   val mailAddress: String
   val mailPassword: String
-
