+++ backend/src/main/scala/com/just/donate/models/Account.scala
@@ -1,6 +1,7 @@
 package com.just.donate.models
 
-import com.just.donate.utils.CollectionUtils.{ map2, updated, updatedReturn }
+import com.just.donate.models.Types.DonationGetter
+import com.just.donate.utils.CollectionUtils.{map2, updated, updatedReturn}
 import com.just.donate.utils.structs.ReservableQueue
 
 case class Account private (
@@ -10,23 +11,6 @@ case class Account private (
 ):
 
   def this(name: String) = this(name, Seq.empty, DonationQueue(name, ReservableQueue(name)))
-
-  def donate(donation: DonationPart): (Boolean, Account) =
-    (true, copy(unboundDonations = unboundDonations.add(donation)))
-
-  // TODO: use error instead of bool
-  def donate(donation: DonationPart, earmarking: String): (Boolean, Account) =
-    def donateRec(queues: Seq[(String, DonationQueue)]): (Boolean, Seq[(String, DonationQueue)]) = queues match
-      case Nil => (false, Nil)
-      case (earmarkingQueue, queue) :: tail =>
-        if earmarking == earmarkingQueue then (true, (earmarkingQueue, queue.add(donation)) +: tail)
-        else
-          val (donated, newQueues) = donateRec(tail)
-          (donated, (earmarkingQueue, queue) +: newQueues)
-
-    val (donated, newBoundDonations) = donateRec(boundDonations)
-    if donated then (true, copy(boundDonations = newBoundDonations))
-    else (false, this)
 
   def removeEarmarking(earmarking: String): Account =
     boundDonations.find(_._1 == earmarking) match
@@ -78,14 +62,6 @@ case class Account private (
       case Some(donationParts) => (donationParts, copy(boundDonations = updatedFrom))
       case None                => throw new IllegalArgumentException(s"Earmarking $earmarking does not exist")
 
-  private def findQueueWithOldestDonation(using donationGetter: DonationGetter): (Option[String], DonationQueue) =
-    val allQueues = (None, unboundDonations) +: boundDonations.map(t => (Some(t._1), t._2))
-    allQueues
-      .filter(_._2.donationQueue._2.nonEmpty)
-      .minByOption(_._2.donationQueue._2.head.value.donation.get.donationDate)
-      // TODO: change to actual error handling
-      .getOrElse(throw new IllegalStateException("No donations available in any queue"))
-
   private[models] def pull(amount: BigDecimal)(using
     donationGetter: DonationGetter
   ): (BigDecimal, DonationPart, Option[String], Account) =
@@ -101,13 +77,38 @@ case class Account private (
 
     (amount - donationPart.head.amount, donationPart.head, earmarking, updatedAccount)
 
-  private[models] def push(donation: DonationPart, earmarking: Option[String]): Account =
-    earmarking match
-      case Some(earmark) => donate(donation, earmark)._2
-      case None          => donate(donation)._2
+  private def findQueueWithOldestDonation(using donationGetter: DonationGetter): (Option[String], DonationQueue) =
+    val allQueues = (None, unboundDonations) +: boundDonations.map(t => (Some(t._1), t._2))
+    allQueues
+      .filter(_._2.donationQueue._2.nonEmpty)
+      .minByOption(_._2.donationQueue._2.head.value.donation.get.donationDate)
+      // TODO: change to actual error handling
+      .getOrElse(throw new IllegalStateException("No donations available in any queue"))
 
   def totalBalance: BigDecimal = totalBalanceUnbound + totalBalanceBound
 
   private def totalBalanceUnbound: BigDecimal = unboundDonations.totalBalance
 
   private def totalBalanceBound: BigDecimal = boundDonations.map(_._2.totalBalance).sum
+
+  private[models] def push(donation: DonationPart, earmarking: Option[String]): Account =
+    earmarking match
+      case Some(earmark) => donate(donation, earmark)._2
+      case None          => donate(donation)._2
+
+  def donate(donation: DonationPart): (Boolean, Account) =
+    (true, copy(unboundDonations = unboundDonations.add(donation)))
+
+  // TODO: use error instead of bool
+  def donate(donation: DonationPart, earmarking: String): (Boolean, Account) =
+    def donateRec(queues: Seq[(String, DonationQueue)]): (Boolean, Seq[(String, DonationQueue)]) = queues match
+      case Nil => (false, Nil)
+      case (earmarkingQueue, queue) :: tail =>
+        if earmarking == earmarkingQueue then (true, (earmarkingQueue, queue.add(donation)) +: tail)
+        else
+          val (donated, newQueues) = donateRec(tail)
+          (donated, (earmarkingQueue, queue) +: newQueues)
+
+    val (donated, newBoundDonations) = donateRec(boundDonations)
+    if donated then (true, copy(boundDonations = newBoundDonations))
+    else (false, this)
