+++ backend/src/main/scala/com/just/donate/models/PaypalIPN.scala
@@ -2,8 +2,7 @@ package com.just.donate.models
 
 import org.mongodb.scala.bson.ObjectId
 
-case class PaypalIPN
-(
+case class PaypalIPN(
   _id: ObjectId = new ObjectId(),
   payload: String
 )
