+++ backend/src/main/scala/com/just/donate/models/Types.scala
@@ -1,0 +1,9 @@
+package com.just.donate.models
+
+/**
+ * Single source of truth for all types used in the application
+ */
+object Types:
+
+  type DonationGetter = String => Option[Donation]
+
