+++ backend/src/test/scala/com/just/donate/models/OrganisationTransferSuite.scala
@@ -3,6 +3,7 @@ package com.just.donate.models
 import com.just.donate.helper.OrganisationHelper.createNewRoots
 import munit.FunSuite
 import com.just.donate.mocks.config.AppConfigMock
+import com.just.donate.models.errors.TransferError
 
 class OrganisationTransferSuite extends FunSuite:
